refactor: split the file functions.ts into modules
