+++ src/app/getUser.ts
@@ -1,0 +1,27 @@
+import { ServerResponse, IncomingMessage } from 'http';
+import { ICodes } from '../interface/Codes';
+import { users } from '../data/database';
+import { v4 as uuidv4, validate as uuidValidate } from 'uuid';
+
+export function getUser(req: IncomingMessage, res: ServerResponse) {
+  try {
+    const urlParts = req.url.split('/');
+    const userId = urlParts[urlParts.length - 1];
+    if (!uuidValidate(userId)) {
+      res.writeHead(ICodes.BAD_REQUEST, { 'Content-Type': 'text/plain' });
+      res.end('Invalid userId');
+      return;
+    }
+    const user = users.find((user) => user.id === userId);
+    if (!user) {
+      res.writeHead(ICodes.NOT_FOUND, { 'Content-Type': 'text/plain' });
+      res.end('User not found');
+      return;
+    }
+    res.writeHead(ICodes.OK, { 'Content-Type': 'application/json' });
+    res.end(JSON.stringify(user));
+  } catch (error) {
+    res.writeHead(ICodes.SERVER_ERROR, { 'Content-Type': 'text/plain' });
+    res.end('500 Internal Server Error');
+  }
+}
